% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb4model.R
\name{feature_splitter}
\alias{feature_splitter}
\title{feature_splitter function Documentation
Splits dataset column names into a tuple of categorical and numerical lists}
\usage{
feature_splitter(x)
}
\value{
a tuple
}
\description{
feature_splitter function Documentation
Splits dataset column names into a tuple of categorical and numerical lists
}
\examples{
feature_splitter(data)
([categorical:],[numerical: ])

}
