% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb4model.R
\name{fit_and_report}
\alias{fit_and_report}
\title{Fit and report}
\usage{
fit_and_report(X, y, Xv, yv, method, m_type = "regression")
}
\arguments{
\item{X}{The features of the training set}

\item{y}{The target of the training set}

\item{Xv}{The feature of the validation set}

\item{yv}{The target of the validation set}

\item{m_type}{The type for calculating error (default = 'regression')}

\item{model}{A machine learning model, can either be classification or regression}
}
\value{
an array of train and validation error
}
\description{
: fits a model and returns the train and validation errors as a list
}
\examples{

}
