% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb4model.R
\name{missing_val}
\alias{missing_val}
\title{Handles missing values}
\usage{
missing_val(df, method)
}
\arguments{
\item{df}{dataframe}

\item{method}{string (''delete' deletes row with missing values, 'avg' replaces missing value with the average, 'last' replaces missing value with the last observation)}
}
\value{
dataframe
}
\description{
Replaces or deletes missing values in a dataframe
}
\examples{
missing_val(df, 'last')

}
