% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb4model.R
\name{ForwardSelection}
\alias{ForwardSelection}
\title{Implement forward feature selection and return data with selected features}
\usage{
ForwardSelection(X, y, min_feat = 1, max_feat = 30)
}
\arguments{
\item{X}{training dataset with features}

\item{y}{training dataset with labels.}

\item{min_feat}{minimum amount of features to select}

\item{max_feat}{maximum amount of features to select}
}
\value{
The dataset with selected features.
}
\description{
Implement forward feature selection and return data with selected features
}
\examples{
add(1, 1)
add(10, 1)

}
