% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_val.R, R/rb4model.R
\name{missing_val}
\alias{missing_val}
\title{Handles missing values}
\usage{
missing_val(df, method)

missing_val(df, method)
}
\arguments{
\item{df}{dataframe with missing values}

\item{method}{string ('delete' removes rows with missing values, 'mean' replaces missing values with averages, 'regression' replaces missings value with fitted values)}
}
\value{
dataframe without missing values

dataframe without missing values
}
\description{
Replaces or deletes missing values in a dataframe

Replaces or deletes missing values in a dataframe
}
\examples{
missing_val(mtcars, 'mean')

missing_val(mtcars, 'mean')

}
